backend/pyproject.toml
[project]
name = "insurance-pricing"
version = "0.1.0"
description = "ML service for predicting insurance premiums based on customer input data with explainability (feature importance, SHAP) and executive summary for the user."
authors = [
    { name = "Viacheslav Danilov", email = "viacheslav.v.danilov@gmail.com" }
]
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.128.0",
    "pydantic>=2.12.5",
    "uvicorn>=0.40.0",
]

[dependency-groups]
dev = [
    "autogluon.tabular>=1.5.0",
    "httpx>=0.28.1",
    "joblib>=1.4.0",
    "matplotlib>=3.9.0",
    "mypy>=1.19.1",
    "numpy>=2.0.0",
    "pandas>=2.2.0",
    "pre-commit>=4.5.1",
    "pytest>=9.0.2",
    "ruff>=0.14.14",
    "scikit-learn>=1.5.0",
    "scipy>=1.14.0",
    "seaborn>=0.13.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/insurance_pricing"]

[tool.mypy]
python_version = "3.13"
exclude = ["venv", ".venv", "tests"]
ignore_missing_imports = true
disallow_untyped_decorators = false
disallow_untyped_calls = false
explicit_package_bases = true

[tool.ruff]
target-version = "py313"
line-length = 88

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
]
ignore = [
    "E501",   # line too long (handled by formatter)
]

[tool.ruff.lint.isort]
known-first-party = ["insurance_pricing"]
